kirkstone image target adaptations
Make bitbaking image target less strict which avoids errors when
packages listed in rootfs manifest cannot be mapped to recipes -- in
such cases we name the recipe as NOT_FOUND instead in the licscantool
output. While this should not typically occur, it helps now to pass
the build with warnings when we are using following declarations
in local.conf as currently suggested in README:

  LICSCAN_SKIP_INHERIT:append = " kernel"
  LICSCAN_SKIP_PN:append = " glibc glibc-locale glibc-mtrace glibc-scripts"

Related to this change we remove now unnecessary
LICSCAN_INCOMPLETE_IMAGE_JSON_WARN_ONLY knob (as the default action
is from here on always 'warn only').

The licscantool script behavior which is also now less strict will now
print also a warning at the end of output in case there were any
warnings during the output of the script to better attract the
attention of the script's user. The warnings from licscantool script
will also be appended to licscan.bbclass generated image specific
licscantool.txt file.

Signed-off-by: Niko Mauno <niko.mauno@vaisala.com>